# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rooms/search": {
      "get": {
        "operationId": "Room_search",
        "parameters": [
          {
            "name": "checkIn",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "checkOut",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-08-03",
              "type": "string"
            }
          },
          {
            "name": "guests",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "roomTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[1-9][0-9]*$",
              "example": "5",
              "type": "string"
            }
          },
          {
            "name": "amenityIds",
            "required": false,
            "in": "query",
            "description": "Repeat this query parameter to require all selected amenities.",
            "schema": {
              "example": [
                "1",
                "2"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "example": "500000.00",
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1500000.00",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RoomSearchSort"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PublicRoomDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      }
    },
    "/api/v1/rooms": {
      "get": {
        "operationId": "Room_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "101",
              "type": "string"
            }
          },
          {
            "name": "roomTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^[1-9][0-9]*$",
              "example": "5",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PublicRoomDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      },
      "post": {
        "operationId": "Room_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 201
                        },
                        "data": {
                          "$ref": "#/components/schemas/RoomDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      }
    },
    "/api/v1/rooms/{id}": {
      "get": {
        "operationId": "Room_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/PublicRoomDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      },
      "patch": {
        "operationId": "Room_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/RoomDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      },
      "delete": {
        "operationId": "Room_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/RoomDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      }
    },
    "/api/v1/rooms/{id}/status": {
      "patch": {
        "operationId": "Room_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/RoomDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/images": {
      "post": {
        "operationId": "Room_createImage",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "isCover": {
                    "type": "boolean",
                    "example": true
                  },
                  "sortOrder": {
                    "type": "integer",
                    "example": 0,
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 201
                        },
                        "data": {
                          "$ref": "#/components/schemas/RoomImageDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Room"
        ]
      }
    },
...
    "/api/v1/management/rooms/{id}": {
      "get": {
        "operationId": "RoomManagement_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/RoomDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Management Rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Homestay Booking Management System API",
    "description": "HTTP contract for the HBMS customer and management applications.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SuccessEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Thanh cong."
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Response payload. The concrete schema is defined per route."
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "f75cf4ac-6958-4d59-8a4f-183abfeaed15"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": [
          "pagination"
        ]
      },
...
      "RoomImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "imageUrl": {
            "type": "string",
            "example": "/uploads/rooms/room-1.jpg"
          },
          "sortOrder": {
            "type": "number",
            "example": 0,
            "minimum": 0
          },
          "isCover": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "imageUrl",
          "sortOrder",
          "isCover"
        ]
      },
      "PublicRoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "roomTypeId": {
            "type": "string",
            "example": "1"
          },
          "roomNumber": {
            "type": "string",
            "example": "101"
          },
          "name": {
            "type": "string",
            "example": "Deluxe Room"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "roomType": {
            "$ref": "#/components/schemas/RoomResponseRoomTypeDto"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomImageDto"
            }
          }
        },
        "required": [
          "id",
          "roomTypeId",
          "roomNumber",
          "name",
          "description",
          "roomType",
          "images"
        ]
      },
      "RoomSearchSort": {
        "type": "string",
        "enum": [
          "RECOMMENDED",
          "PRICE_ASC",
          "PRICE_DESC",
          "POPULARITY",
          "NEWEST"
        ]
      },
...
      "RoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "roomTypeId": {
            "type": "string",
            "example": "1"
          },
          "roomNumber": {
            "type": "string",
            "example": "101"
          },
          "name": {
            "type": "string",
            "example": "Deluxe 101"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              }
            ]
          },
          "roomType": {
            "$ref": "#/components/schemas/RoomResponseRoomTypeDto"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomImageDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "roomTypeId",
          "roomNumber",
          "name",
          "description",
          "status",
          "roomType",
          "images",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "roomTypeId": {
            "type": "string",
            "example": "5",
            "pattern": "^[1-9][0-9]*$"
          },
          "roomNumber": {
            "type": "string",
            "example": "101",
            "maxLength": 50
          },
          "name": {
            "type": "string",
            "example": "Phòng 101",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Phòng tầng một, gần khu vườn.",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "OCCUPIED",
              "CLEANING",
              "MAINTENANCE",
              "HIDDEN"
            ],
            "example": "READY"
          }
        },
        "required": [
          "roomTypeId",
          "roomNumber",
          "name"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "roomTypeId": {
            "type": "string",
            "example": "5",
            "pattern": "^[1-9][0-9]*$"
          },
          "roomNumber": {
            "type": "string",
            "example": "101",
            "maxLength": 50
          },
          "name": {
            "type": "string",
            "example": "Phòng 101 hướng vườn",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Phòng tầng một, gần khu vườn.",
            "nullable": true
          }
        }
      },
      "UpdateRoomStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "OCCUPIED",
              "CLEANING",
              "MAINTENANCE",
              "HIDDEN"
            ],
            "example": "MAINTENANCE"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}